+++ __test__/helper/app.js
@@ -4,7 +4,18 @@ const app = require("../../src/server");
 const { response } = require("../../src/server");
 const request = supertest(app)
 const assert = require('assert');
+const assertEquals = require()
 
+const characters = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
 
+function generateString(length) {
+    let result = ' ';
+    const charactersLength = characters.length;
+    for (let i = 0; i < length; i++) {
+        result += characters.charAt(Math.floor(Math.random() * charactersLength));
+    }
 
-module.exports = { request, assert }
+    return result;
+}
+
+module.exports = { request, assert, generateString }
