+++ __test__/room.spec.js
@@ -1,7 +1,9 @@
-const { request, assert } = require("./helper/app");
+const { request, assert, generateString } = require("./helper/app");
 const { createRoom } = require("./helper/room_test_data");
 const { createUser } = require('./helper/Student_test_data')
 const dbHandler = require('./helper/db-handler')
+
+// Room testing
 
 /**
  * Connect to a new in-memory database before running any tests.
@@ -19,18 +21,6 @@ afterAll(async () => {
 });
 
 describe("testing room api routes", () => {
-
-    const characters = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
-
-    function generateString(length) {
-        let result = ' ';
-        const charactersLength = characters.length;
-        for (let i = 0; i < length; i++) {
-            result += characters.charAt(Math.floor(Math.random() * charactersLength));
-        }
-
-        return result;
-    }
 
     let user1 = createUser();
     let user2 = createUser();
