+++ __test__/student.spec.js
@@ -2,6 +2,8 @@ const { request, assert } = require("./helper/app");
 const { createUser } = require('./helper/Student_test_data')
 const dbHandler = require('./helper/db-handler')
 
+// Student testing
+
 /**
  * Connect to a new in-memory database before running any tests.
  */
